# Conflicts:
#	src/main/resources/static/app/views/create-meeting.html

+++ src/main/java/za/co/bsg/services/MeetingDataService.java
@@ -13,7 +13,7 @@ public class MeetingDataService {
     @Autowired
     MeetingRepository meetingRepository;
 
-    public List<Meeting> retrieveAllByUserId(int userId){
+    public List<Meeting> retrieveAllByUserId(long userId){
         return meetingRepository.findByCreatedBy(userId);
     }
